requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
gunicorn==22.0.0
requests==2.32.3
python-dotenv==1.0.1
Pillow==10.4.0
